Modified the AppendToDat method, now takes a FileIndex object instead of a int (array index). Other excel files updated, with unused columns hidden (made private) and other various updates.

+++ trunk/source/ExcelFile.cs
@@ -548,10 +548,10 @@ namespace Reanimator
                 {
                     case 2:
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur)); // int value between 270-370
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur)); // int value between 270-370
                         cur += sizeof(int);
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur)); // 0
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur)); // 0
                         cur += sizeof(int);
                         csv.Write(",");
                         break;
@@ -566,7 +566,7 @@ namespace Reanimator
                     case 527:
                     case 700:
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur)); // int value
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur)); // int value
                         cur += sizeof(int);
                         csv.Write(",");
                         break;
@@ -603,7 +603,7 @@ namespace Reanimator
                     case 687:
                     case 688:
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur)); // int bitmask
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur)); // int bitmask
                         cur += sizeof(int);
                         csv.Write(",");
                         break;
@@ -611,47 +611,47 @@ namespace Reanimator
                     // functions
                     case 707:
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur)); // Boolean? 1/0
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur)); // Boolean? 1/0
                         cur += sizeof(int);
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur));  // 714
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur));  // 714
                         cur += sizeof(int);
                         csv.Write(",");
                         break;
                     // random range?
                     case 708: // 26,3,26
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur)); // 26
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur)); // 26
                         cur += sizeof(int);
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur));  // value
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur));  // value
                         cur += sizeof(int);
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur)); // 3
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur)); // 3
                         cur += sizeof(int);
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur));  // value
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur));  // value
                         cur += sizeof(int);
                         csv.Write(",");
                         break;
                     case 710: // 26,26,3
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur)); // 26
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur)); // 26
                         cur += sizeof(int);
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur));  // value
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur));  // value
                         cur += sizeof(int);
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur)); // 26
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur)); // 26
                         cur += sizeof(int);
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur));  // value
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur));  // value
                         cur += sizeof(int);
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur)); // 3
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur)); // 3
                         cur += sizeof(int);
                         csv.Write(",");
-                        csv.Write(FileTools.ByteArrayToInt32(_dataBlock, cur)); // value
+                        csv.Write((uint)FileTools.ByteArrayToInt32(_dataBlock, cur)); // value
                         cur += sizeof(int);
                         csv.Write(",");
                         break;
@@ -926,7 +926,7 @@ namespace Reanimator
 
                                 foreach (string part in explode)
                                 {
-                                    FileTools.WriteToBuffer(ref intBytes, ref intByteCount, int.Parse(part));
+                                    FileTools.WriteToBuffer(ref intBytes, ref intByteCount, uint.Parse(part));
                                 }
                             }
                         }
